+++ datarobot_provider/example_dags/deprecated/datarobot_datetime_partitioning_pipeline_dag.py
