+++ tests/unit/dags/pipline/test_datarobot_pipeline_dag.py
@@ -7,7 +7,7 @@
 # Released under the terms of DataRobot Tool and Utility Agreement.
 import pytest
 
-from datarobot_provider.example_dags.datarobot_pipeline_dag import datarobot_pipeline
+from datarobot_provider.example_dags.deprecated.datarobot_pipeline_dag import datarobot_pipeline
 
 
 def test_dag_loaded(dagbag):
