+++ tests/unit/dags/project/test_create_project_from_dataset_version_dag.py
@@ -7,7 +7,7 @@
 # Released under the terms of DataRobot Tool and Utility Agreement.
 import pytest
 
-from datarobot_provider.example_dags.datarobot_create_project_from_dataset_version_dag import (
+from datarobot_provider.example_dags.deprecated.datarobot_create_project_from_dataset_version_dag import (
     create_project_from_dataset_version,
 )
